refactor: move downloadUpstreamHermesTarball to CI script
downloadUpstreamHermesTarball is only called from CI, not by any
library consumer. Move it (and its computeNightlyTarballURL dependency)
to .github/scripts/resolve-hermes.mts where it belongs.

microsoft-resolveHermes.js is now a pure library with only the
functions imported by hermes.js and reactNativeDependencies.js.
